split api.go into different files
