+++ pkg/grpc/default_api/api.go
@@ -19,10 +19,7 @@ package default_api
 import (
 	"context"
 	"errors"
-	"fmt"
 	"github.com/dapr/components-contrib/secretstores"
-	"io"
-	"strings"
 	"sync"
 
 	"github.com/dapr/components-contrib/bindings"
@@ -32,17 +29,12 @@ import (
 	dapr_v1pb "mosn.io/layotto/pkg/grpc/dapr/proto/runtime/v1"
 	mgrpc "mosn.io/mosn/pkg/filter/network/grpc"
 
-	"github.com/dapr/components-contrib/state"
-	"github.com/golang/protobuf/ptypes/empty"
-	"mosn.io/layotto/components/file"
-	"mosn.io/pkg/utils"
-
 	"github.com/dapr/components-contrib/pubsub"
+	"github.com/dapr/components-contrib/state"
 	jsoniter "github.com/json-iterator/go"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/emptypb"
+	"mosn.io/layotto/components/file"
 
 	"mosn.io/layotto/components/configstores"
 	"mosn.io/layotto/components/hello"
@@ -263,384 +255,4 @@ func (a *api) InvokeService(ctx context.Context, in *runtimev1pb.InvokeServiceRe
 		Data:        daprResp.Data,
 		ContentType: daprResp.ContentType,
 	}, nil
-}
-
-// GetConfiguration gets configuration from configuration store.
-func (a *api) GetConfiguration(ctx context.Context, req *runtimev1pb.GetConfigurationRequest) (*runtimev1pb.GetConfigurationResponse, error) {
-	resp := &runtimev1pb.GetConfigurationResponse{}
-	// check store type supported or not
-	store, ok := a.configStores[req.StoreName]
-	if !ok {
-		return nil, errors.New(fmt.Sprintf("configure store [%+v] don't support now", req.StoreName))
-	}
-	//here protect user use space for sting, eg: " ", "de fault"
-	if strings.ReplaceAll(req.Group, " ", "") == "" {
-		req.Group = store.GetDefaultGroup()
-	}
-	if strings.ReplaceAll(req.Label, " ", "") == "" {
-		req.Label = store.GetDefaultLabel()
-	}
-	items, err := store.Get(ctx, &configstores.GetRequest{AppId: req.AppId, Group: req.Group, Label: req.Label, Keys: req.Keys, Metadata: req.Metadata})
-	if err != nil {
-		return nil, errors.New(fmt.Sprintf("get configuration failed with error: %+v", err))
-	}
-	for _, item := range items {
-		resp.Items = append(resp.Items, &runtimev1pb.ConfigurationItem{Group: item.Group, Label: item.Label, Key: item.Key, Content: item.Content, Tags: item.Tags, Metadata: item.Metadata})
-	}
-	return resp, err
-}
-
-// SaveConfiguration saves configuration into configuration store.
-func (a *api) SaveConfiguration(ctx context.Context, req *runtimev1pb.SaveConfigurationRequest) (*emptypb.Empty, error) {
-	store, ok := a.configStores[req.StoreName]
-	if !ok {
-		return nil, errors.New(fmt.Sprintf("configure store [%+v] don't support now", req.StoreName))
-	}
-	setReq := &configstores.SetRequest{}
-	setReq.AppId = req.AppId
-	setReq.StoreName = req.StoreName
-
-	for index, item := range req.Items {
-		if strings.ReplaceAll(item.Group, " ", "") == "" {
-			req.Items[index].Group = store.GetDefaultGroup()
-		}
-		if strings.ReplaceAll(item.Label, " ", "") == "" {
-			req.Items[index].Label = store.GetDefaultLabel()
-		}
-		setReq.Items = append(setReq.Items, &configstores.ConfigurationItem{Group: item.Group, Label: item.Label, Key: item.Key, Content: item.Content, Tags: item.Tags, Metadata: item.Metadata})
-	}
-	err := store.Set(ctx, setReq)
-	return &emptypb.Empty{}, err
-}
-
-// DeleteConfiguration deletes configuration from configuration store.
-func (a *api) DeleteConfiguration(ctx context.Context, req *runtimev1pb.DeleteConfigurationRequest) (*emptypb.Empty, error) {
-	store, ok := a.configStores[req.StoreName]
-	if !ok {
-		return nil, errors.New(fmt.Sprintf("configure store [%+v] don't support now", req.StoreName))
-	}
-	if strings.ReplaceAll(req.Group, " ", "") == "" {
-		req.Group = store.GetDefaultGroup()
-	}
-	if strings.ReplaceAll(req.Label, " ", "") == "" {
-		req.Label = store.GetDefaultLabel()
-	}
-	err := store.Delete(ctx, &configstores.DeleteRequest{AppId: req.AppId, Group: req.Group, Label: req.Label, Keys: req.Keys, Metadata: req.Metadata})
-	return &emptypb.Empty{}, err
-}
-
-// SubscribeConfiguration gets configuration from configuration store and subscribe the updates.
-func (a *api) SubscribeConfiguration(sub runtimev1pb.Runtime_SubscribeConfigurationServer) error {
-	wg := sync.WaitGroup{}
-	wg.Add(2)
-	var subErr error
-	respCh := make(chan *configstores.SubscribeResp)
-	recvExitCh := make(chan struct{})
-	subscribedStore := make([]configstores.Store, 0, 1)
-	// TODO currently this goroutine model is error-prone,and it should be refactored after new version of configuration API being accepted
-	// 1. start a reader goroutine
-	utils.GoWithRecover(func() {
-		defer wg.Done()
-		for {
-			// 1.1. read stream
-			req, err := sub.Recv()
-			// 1.2. if an error happens,stop all the subscribers
-			if err != nil {
-				log.DefaultLogger.Errorf("occur error in subscribe, err: %+v", err)
-				// stop all the subscribers
-				for _, store := range subscribedStore {
-					// TODO this method will stop subscribers created by other connections.Should be refactored
-					store.StopSubscribe()
-				}
-				subErr = err
-				// stop writer goroutine
-				close(recvExitCh)
-				return
-			}
-			// 1.3. else find the component and delegate to it
-			store, ok := a.configStores[req.StoreName]
-			// 1.3.1. stop if StoreName is not supported
-			if !ok {
-				log.DefaultLogger.Errorf("configure store [%+v] don't support now", req.StoreName)
-				// stop all the subscribers
-				for _, store := range subscribedStore {
-					store.StopSubscribe()
-				}
-				subErr = errors.New(fmt.Sprintf("configure store [%+v] don't support now", req.StoreName))
-				// stop writer goroutine
-				close(recvExitCh)
-				return
-			}
-			// 1.3.2. use default settings if blank
-			if strings.ReplaceAll(req.Group, " ", "") == "" {
-				req.Group = store.GetDefaultGroup()
-			}
-			if strings.ReplaceAll(req.Label, " ", "") == "" {
-				req.Label = store.GetDefaultLabel()
-			}
-			// 1.3.3. delegate to the component
-			store.Subscribe(&configstores.SubscribeReq{AppId: req.AppId, Group: req.Group, Label: req.Label, Keys: req.Keys, Metadata: req.Metadata}, respCh)
-			subscribedStore = append(subscribedStore, store)
-		}
-	}, nil)
-	// 2. start a writer goroutine
-	utils.GoWithRecover(func() {
-		defer wg.Done()
-		for {
-			select {
-			// read response from components
-			case resp, ok := <-respCh:
-				if !ok {
-					return
-				}
-				items := make([]*runtimev1pb.ConfigurationItem, 0, 10)
-				for _, item := range resp.Items {
-					items = append(items, &runtimev1pb.ConfigurationItem{Group: item.Group, Label: item.Label, Key: item.Key, Content: item.Content, Tags: item.Tags, Metadata: item.Metadata})
-				}
-				// write to response stream
-				sub.Send(&runtimev1pb.SubscribeConfigurationResponse{StoreName: resp.StoreName, AppId: resp.StoreName, Items: items})
-			//	read exit signal
-			case <-recvExitCh:
-				return
-			}
-		}
-	}, nil)
-	wg.Wait()
-	log.DefaultLogger.Warnf("subscribe gorountine exit")
-	return subErr
-}
-
-func (a *api) GetFile(req *runtimev1pb.GetFileRequest, stream runtimev1pb.Runtime_GetFileServer) error {
-	if a.fileOps[req.StoreName] == nil {
-		return status.Errorf(codes.InvalidArgument, "not supported store type: %+v", req.StoreName)
-	}
-	if req.Metadata == nil {
-		req.Metadata = make(map[string]string)
-	}
-	st := &file.GetFileStu{FileName: req.Name, Metadata: req.Metadata}
-	data, err := a.fileOps[req.StoreName].Get(stream.Context(), st)
-	if err != nil {
-		return status.Errorf(codes.Internal, "get file fail,err: %+v", err)
-	}
-
-	buffsPtr := bytesPool.Get().(*[]byte)
-	buf := *buffsPtr
-	if len(buf) == 0 {
-		buf = make([]byte, 102400, 102400)
-	}
-	defer func() {
-		data.Close()
-		*buffsPtr = buf
-		bytesPool.Put(buffsPtr)
-	}()
-
-	for {
-		length, err := data.Read(buf)
-		if err != nil && err != io.EOF {
-			log.DefaultLogger.Warnf("get file fail, err: %+v", err)
-			return status.Errorf(codes.Internal, "get file fail,err: %+v", err)
-		}
-		if err == nil || (err == io.EOF && length != 0) {
-			resp := &runtimev1pb.GetFileResponse{Data: buf[:length]}
-			if err = stream.Send(resp); err != nil {
-				return status.Errorf(codes.Internal, "send file data fail,err: %+v", err)
-			}
-		}
-		if err == io.EOF {
-			return nil
-		}
-	}
-}
-
-type putObjectStreamReader struct {
-	data   []byte
-	server runtimev1pb.Runtime_PutFileServer
-}
-
-func newPutObjectStreamReader(data []byte, server runtimev1pb.Runtime_PutFileServer) *putObjectStreamReader {
-	return &putObjectStreamReader{data: data, server: server}
-}
-
-func (r *putObjectStreamReader) Read(p []byte) (int, error) {
-	var count int
-	total := len(p)
-	for {
-		if len(r.data) > 0 {
-			n := copy(p[count:], r.data)
-			r.data = r.data[n:]
-			count += n
-			if count == total {
-				return count, nil
-			}
-		}
-		req, err := r.server.Recv()
-		if err != nil {
-			if err != io.EOF {
-				log.DefaultLogger.Errorf("recv data from grpc stream fail, err:%+v", err)
-			}
-			return count, err
-		}
-		r.data = req.Data
-	}
-}
-
-func (a *api) PutFile(stream runtimev1pb.Runtime_PutFileServer) error {
-	req, err := stream.Recv()
-	if err != nil {
-		//if client send eof error directly, return nil
-		if err == io.EOF {
-			return nil
-		}
-		return status.Errorf(codes.Internal, "receive file data fail: err: %+v", err)
-	}
-
-	if a.fileOps[req.StoreName] == nil {
-		return status.Errorf(codes.InvalidArgument, "not support store type: %+v", req.StoreName)
-	}
-	fileReader := newPutObjectStreamReader(req.Data, stream)
-	if req.Metadata == nil {
-		req.Metadata = make(map[string]string)
-	}
-	st := &file.PutFileStu{DataStream: fileReader, FileName: req.Name, Metadata: req.Metadata}
-	if err = a.fileOps[req.StoreName].Put(stream.Context(), st); err != nil {
-		return status.Errorf(codes.Internal, err.Error())
-	}
-	stream.SendAndClose(&empty.Empty{})
-	return nil
-}
-
-//ListFile list all files
-func (a *api) ListFile(ctx context.Context, in *runtimev1pb.ListFileRequest) (*runtimev1pb.ListFileResp, error) {
-	if in.Request == nil {
-		return nil, status.Errorf(codes.InvalidArgument, "request can't be nil")
-	}
-	if in.Request.Metadata == nil {
-		in.Request.Metadata = make(map[string]string)
-	}
-
-	if a.fileOps[in.Request.StoreName] == nil {
-		return nil, status.Errorf(codes.InvalidArgument, "not support store type: %+v", in.Request.StoreName)
-	}
-	resp, err := a.fileOps[in.Request.StoreName].List(ctx, &file.ListRequest{DirectoryName: in.Request.Name, PageSize: in.PageSize, Marker: in.Marker, Metadata: in.Request.Metadata})
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, err.Error())
-	}
-	files := make([]*runtimev1pb.FileInfo, 0)
-	for _, v := range resp.Files {
-		file := &runtimev1pb.FileInfo{}
-		file.FileName = v.FileName
-		file.LastModified = v.LastModified
-		file.Size = v.Size
-		file.Metadata = v.Meta
-		files = append(files, file)
-	}
-	return &runtimev1pb.ListFileResp{Files: files, Marker: resp.Marker, IsTruncated: resp.IsTruncated}, nil
-}
-
-//DelFile delete specific file
-func (a *api) DelFile(ctx context.Context, in *runtimev1pb.DelFileRequest) (*emptypb.Empty, error) {
-	errCode := codes.Internal
-	if in.Request == nil {
-		return nil, status.Errorf(codes.InvalidArgument, "request can't be nil")
-	}
-	if in.Request.Metadata == nil {
-		in.Request.Metadata = make(map[string]string)
-	}
-	if a.fileOps[in.Request.StoreName] == nil {
-		return nil, status.Errorf(codes.InvalidArgument, "not support store type: %+v", in.Request.StoreName)
-	}
-	err := a.fileOps[in.Request.StoreName].Del(ctx, &file.DelRequest{FileName: in.Request.Name, Metadata: in.Request.Metadata})
-	if err != nil {
-		if code, ok := FileErrMap2GrpcErr[err]; ok {
-			errCode = code
-		}
-		return nil, status.Errorf(errCode, err.Error())
-	}
-	return &emptypb.Empty{}, nil
-}
-
-//GetFileMeta get meta of file
-func (a *api) GetFileMeta(ctx context.Context, in *runtimev1pb.GetFileMetaRequest) (*runtimev1pb.GetFileMetaResponse, error) {
-	errCode := codes.Internal
-	if in.Request == nil {
-		return nil, status.Errorf(codes.InvalidArgument, "request can't be nil")
-	}
-	if in.Request.Metadata == nil {
-		in.Request.Metadata = make(map[string]string)
-	}
-	if a.fileOps[in.Request.StoreName] == nil {
-		return nil, status.Errorf(codes.InvalidArgument, "not support store type: %+v", in.Request.StoreName)
-	}
-	resp, err := a.fileOps[in.Request.StoreName].Stat(ctx, &file.FileMetaRequest{FileName: in.Request.Name, Metadata: in.Request.Metadata})
-	if err != nil {
-		if code, ok := FileErrMap2GrpcErr[err]; ok {
-			errCode = code
-		}
-		return nil, status.Errorf(errCode, err.Error())
-	}
-	meta := &runtimev1pb.FileMeta{}
-	meta.Metadata = make(map[string]*runtimev1pb.FileMetaValue)
-	for k, v := range resp.Metadata {
-		meta.Metadata[k] = &runtimev1pb.FileMetaValue{}
-		meta.Metadata[k].Value = append(meta.Metadata[k].Value, v...)
-	}
-	return &runtimev1pb.GetFileMetaResponse{Size: resp.Size, LastModified: resp.LastModified, Response: meta}, nil
-}
-
-func (a *api) InvokeBinding(ctx context.Context, in *runtimev1pb.InvokeBindingRequest) (*runtimev1pb.InvokeBindingResponse, error) {
-	daprResp, err := a.daprAPI.InvokeBinding(ctx, &dapr_v1pb.InvokeBindingRequest{
-		Name:      in.Name,
-		Data:      in.Data,
-		Metadata:  in.Metadata,
-		Operation: in.Operation,
-	})
-	if err != nil {
-		return &runtimev1pb.InvokeBindingResponse{}, err
-	}
-	return &runtimev1pb.InvokeBindingResponse{
-		Data:     daprResp.Data,
-		Metadata: daprResp.Metadata,
-	}, nil
-}
-
-func (a *api) GetSecret(ctx context.Context, in *runtimev1pb.GetSecretRequest) (*runtimev1pb.GetSecretResponse, error) {
-	daprResp, err := a.daprAPI.GetSecret(ctx, &dapr_v1pb.GetSecretRequest{
-		StoreName: in.StoreName,
-		Key:       in.Key,
-		Metadata:  in.Metadata,
-	})
-	if err != nil {
-		return &runtimev1pb.GetSecretResponse{}, err
-	}
-	return &runtimev1pb.GetSecretResponse{Data: daprResp.Data}, nil
-}
-
-func (a *api) GetBulkSecret(ctx context.Context, in *runtimev1pb.GetBulkSecretRequest) (*runtimev1pb.GetBulkSecretResponse, error) {
-	daprResp, err := a.daprAPI.GetBulkSecret(ctx, &dapr_v1pb.GetBulkSecretRequest{
-		StoreName: in.StoreName,
-		Metadata:  in.Metadata,
-	})
-	if err != nil {
-		return &runtimev1pb.GetBulkSecretResponse{}, err
-	}
-	return &runtimev1pb.GetBulkSecretResponse{
-		Data: convertSecretResponseMap(daprResp.Data),
-	}, nil
-}
-
-func convertSecretResponseMap(data map[string]*dapr_v1pb.SecretResponse) map[string]*runtimev1pb.SecretResponse {
-	if data == nil {
-		return nil
-	}
-	result := make(map[string]*runtimev1pb.SecretResponse)
-	for k, v := range data {
-		var converted *runtimev1pb.SecretResponse
-		if v != nil {
-			converted = &runtimev1pb.SecretResponse{
-				Secrets: v.Secrets,
-			}
-		}
-		result[k] = converted
-	}
-	return result
 }
