+++ pkg/grpc/default_api/api_binding_test.go
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2021 Layotto Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package default_api
+
+import (
+	"context"
+	"errors"
+	"github.com/dapr/components-contrib/bindings"
+	"github.com/phayes/freeport"
+	"github.com/stretchr/testify/assert"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
+	"testing"
+)
+
+func TestInvokeBinding(t *testing.T) {
+	port, _ := freeport.GetFreePort()
+	srv := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil,
+		func(name string, req *bindings.InvokeRequest) (*bindings.InvokeResponse, error) {
+			if name == "error-binding" {
+				return nil, errors.New("error when invoke binding")
+			}
+			return &bindings.InvokeResponse{Data: []byte("ok")}, nil
+		}, nil)
+	server := startTestServerAPI(port, srv)
+	defer server.Stop()
+
+	clientConn := createTestClient(port)
+	defer clientConn.Close()
+
+	client := runtimev1pb.NewRuntimeClient(clientConn)
+	_, err := client.InvokeBinding(context.Background(), &runtimev1pb.InvokeBindingRequest{})
+	assert.Nil(t, err)
+	_, err = client.InvokeBinding(context.Background(), &runtimev1pb.InvokeBindingRequest{Name: "error-binding"})
+	assert.Equal(t, codes.Internal, status.Code(err))
+}
