+++ pkg/grpc/default_api/api_lock_test.go
@@ -14,10 +14,14 @@
 package default_api
 
 import (
+	"context"
+	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
-	"mosn.io/layotto/components/lock"
-	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
 	"testing"
+
+	"mosn.io/layotto/components/lock"
+	mock_lock "mosn.io/layotto/pkg/mock/components/lock"
+	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
 )
 
 func TestTryLockRequest2ComponentRequest(t *testing.T) {
@@ -60,4 +64,148 @@ func TestUnlockComp2GrpcResponse(t *testing.T) {
 	assert.True(t, resp.Status == runtimev1pb.UnlockResponse_SUCCESS)
 	resp2 := UnlockComp2GrpcResponse(nil)
 	assert.NotNil(t, resp2)
+}
+
+func TestTryLock(t *testing.T) {
+	t.Run("lock store not configured", func(t *testing.T) {
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil, nil, nil)
+		req := &runtimev1pb.TryLockRequest{
+			StoreName: "abc",
+		}
+		_, err := api.TryLock(context.Background(), req)
+		assert.Equal(t, "rpc error: code = FailedPrecondition desc = lock store is not configured", err.Error())
+	})
+
+	t.Run("resourceid empty", func(t *testing.T) {
+		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
+		req := &runtimev1pb.TryLockRequest{
+			StoreName: "abc",
+		}
+		_, err := api.TryLock(context.Background(), req)
+		assert.Equal(t, "rpc error: code = InvalidArgument desc = ResourceId is empty in lock store abc", err.Error())
+	})
+
+	t.Run("lock owner empty", func(t *testing.T) {
+		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
+		req := &runtimev1pb.TryLockRequest{
+			StoreName:  "abc",
+			ResourceId: "resource",
+		}
+		_, err := api.TryLock(context.Background(), req)
+		assert.Equal(t, "rpc error: code = InvalidArgument desc = LockOwner is empty in lock store abc", err.Error())
+	})
+
+	t.Run("lock expire is not positive", func(t *testing.T) {
+		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
+		req := &runtimev1pb.TryLockRequest{
+			StoreName:  "abc",
+			ResourceId: "resource",
+			LockOwner:  "owner",
+		}
+		_, err := api.TryLock(context.Background(), req)
+		assert.Equal(t, "rpc error: code = InvalidArgument desc = Expire is not positive in lock store abc", err.Error())
+	})
+
+	t.Run("lock store not found", func(t *testing.T) {
+		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
+		req := &runtimev1pb.TryLockRequest{
+			StoreName:  "abc",
+			ResourceId: "resource",
+			LockOwner:  "owner",
+			Expire:     1,
+		}
+		_, err := api.TryLock(context.Background(), req)
+		assert.Equal(t, "rpc error: code = InvalidArgument desc = lock store abc not found", err.Error())
+	})
+
+	t.Run("normal", func(t *testing.T) {
+		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
+		mockLockStore.EXPECT().TryLock(gomock.Any()).DoAndReturn(func(req *lock.TryLockRequest) (*lock.TryLockResponse, error) {
+			assert.Equal(t, "lock|||resource", req.ResourceId)
+			assert.Equal(t, "owner", req.LockOwner)
+			assert.Equal(t, int32(1), req.Expire)
+			return &lock.TryLockResponse{
+				Success: true,
+			}, nil
+		})
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
+		req := &runtimev1pb.TryLockRequest{
+			StoreName:  "mock",
+			ResourceId: "resource",
+			LockOwner:  "owner",
+			Expire:     1,
+		}
+		resp, err := api.TryLock(context.Background(), req)
+		assert.Nil(t, err)
+		assert.Equal(t, true, resp.Success)
+	})
+
+}
+
+func TestUnlock(t *testing.T) {
+	t.Run("lock store not configured", func(t *testing.T) {
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil, nil, nil)
+		req := &runtimev1pb.UnlockRequest{
+			StoreName: "abc",
+		}
+		_, err := api.Unlock(context.Background(), req)
+		assert.Equal(t, "rpc error: code = FailedPrecondition desc = lock store is not configured", err.Error())
+	})
+
+	t.Run("resourceid empty", func(t *testing.T) {
+		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
+		req := &runtimev1pb.UnlockRequest{
+			StoreName: "abc",
+		}
+		_, err := api.Unlock(context.Background(), req)
+		assert.Equal(t, "rpc error: code = InvalidArgument desc = ResourceId is empty in lock store abc", err.Error())
+	})
+
+	t.Run("lock owner empty", func(t *testing.T) {
+		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
+		req := &runtimev1pb.UnlockRequest{
+			StoreName:  "abc",
+			ResourceId: "resource",
+		}
+		_, err := api.Unlock(context.Background(), req)
+		assert.Equal(t, "rpc error: code = InvalidArgument desc = LockOwner is empty in lock store abc", err.Error())
+	})
+
+	t.Run("lock store not found", func(t *testing.T) {
+		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
+		req := &runtimev1pb.UnlockRequest{
+			StoreName:  "abc",
+			ResourceId: "resource",
+			LockOwner:  "owner",
+		}
+		_, err := api.Unlock(context.Background(), req)
+		assert.Equal(t, "rpc error: code = InvalidArgument desc = lock store abc not found", err.Error())
+	})
+
+	t.Run("normal", func(t *testing.T) {
+		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
+		mockLockStore.EXPECT().Unlock(gomock.Any()).DoAndReturn(func(req *lock.UnlockRequest) (*lock.UnlockResponse, error) {
+			assert.Equal(t, "lock|||resource", req.ResourceId)
+			assert.Equal(t, "owner", req.LockOwner)
+			return &lock.UnlockResponse{
+				Status: lock.SUCCESS,
+			}, nil
+		})
+		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
+		req := &runtimev1pb.UnlockRequest{
+			StoreName:  "mock",
+			ResourceId: "resource",
+			LockOwner:  "owner",
+		}
+		resp, err := api.Unlock(context.Background(), req)
+		assert.Nil(t, err)
+		assert.Equal(t, runtimev1pb.UnlockResponse_SUCCESS, resp.Status)
+	})
 }
