+++ pkg/grpc/default_api/api_test.go
@@ -19,34 +19,19 @@ package default_api
 import (
 	"context"
 	"fmt"
-	"github.com/dapr/components-contrib/secretstores"
-	moke_secret "mosn.io/layotto/pkg/mock/components/secret"
-
-	"github.com/dapr/components-contrib/bindings"
 	"github.com/golang/mock/gomock"
-	"github.com/phayes/freeport"
 	"github.com/stretchr/testify/assert"
-	"google.golang.org/grpc/codes"
 	l8grpc "mosn.io/layotto/pkg/grpc"
-	"mosn.io/layotto/pkg/mock/runtime"
 	"net"
 	"testing"
 
-	"errors"
-	"google.golang.org/grpc/status"
-	"io"
-	"mosn.io/layotto/components/configstores"
 	"mosn.io/layotto/components/hello"
-	"mosn.io/layotto/components/lock"
 	"mosn.io/layotto/components/rpc"
 	"mosn.io/layotto/pkg/mock"
 	mock_invoker "mosn.io/layotto/pkg/mock/components/invoker"
-	mock_lock "mosn.io/layotto/pkg/mock/components/lock"
 	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
 
 	"time"
-
-	"mosn.io/layotto/components/file"
 
 	"github.com/golang/protobuf/ptypes/any"
 	tmock "github.com/stretchr/testify/mock"
@@ -59,20 +44,6 @@ const (
 	maxGRPCServerUptime = 100 * time.Millisecond
 	testGRPCServerPort  = 19887
 )
-
-type MockGrpcServer struct {
-	err error
-	req *runtimev1pb.SubscribeConfigurationRequest
-	grpc.ServerStream
-}
-
-func (m *MockGrpcServer) Send(res *runtimev1pb.SubscribeConfigurationResponse) error {
-	return nil
-}
-
-func (m *MockGrpcServer) Recv() (*runtimev1pb.SubscribeConfigurationRequest, error) {
-	return m.req, m.err
-}
 
 type mockGRPCAPI struct {
 	API
@@ -151,113 +122,6 @@ func startTestRuntimeAPIServer(port int, testAPIServer API) *grpc.Server {
 	return server
 }
 
-func TestGetConfiguration(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	mockConfigStore := mock.NewMockStore(ctrl)
-	api := NewAPI("", nil, map[string]configstores.Store{"mock": mockConfigStore}, nil, nil, nil, nil, nil, nil, nil, nil)
-	mockConfigStore.EXPECT().Get(gomock.Any(), gomock.Any()).Return([]*configstores.ConfigurationItem{
-		{Key: "sofa", Content: "sofa1"},
-	}, nil).Times(1)
-	res, err := api.GetConfiguration(context.Background(), &runtimev1pb.GetConfigurationRequest{StoreName: "mock", AppId: "mosn", Keys: []string{"sofa"}})
-	assert.Nil(t, err)
-	assert.Equal(t, res.Items[0].Key, "sofa")
-	assert.Equal(t, res.Items[0].Content, "sofa1")
-	_, err = api.GetConfiguration(context.Background(), &runtimev1pb.GetConfigurationRequest{StoreName: "etcd", AppId: "mosn", Keys: []string{"sofa"}})
-	assert.Equal(t, err.Error(), "configure store [etcd] don't support now")
-
-}
-
-func TestSaveConfiguration(t *testing.T) {
-	t.Run("normal", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockConfigStore := mock.NewMockStore(ctrl)
-		mockConfigStore.EXPECT().Set(gomock.Any(), gomock.Any()).DoAndReturn(func(ctx context.Context, req *configstores.SetRequest) error {
-			assert.Equal(t, "appid", req.AppId)
-			assert.Equal(t, "mock", req.StoreName)
-			assert.Equal(t, 1, len(req.Items))
-			return nil
-		})
-		req := &runtimev1pb.SaveConfigurationRequest{
-			StoreName: "mock",
-			AppId:     "appid",
-			Items: []*runtimev1pb.ConfigurationItem{
-				{
-					Key:      "key",
-					Content:  "value",
-					Group:    "  ",
-					Label:    "  ",
-					Tags:     nil,
-					Metadata: nil,
-				},
-			},
-			Metadata: nil,
-		}
-		api := NewAPI("", nil, map[string]configstores.Store{"mock": mockConfigStore}, nil, nil, nil, nil, nil, nil, nil, nil)
-		_, err := api.SaveConfiguration(context.Background(), req)
-		assert.Nil(t, err)
-	})
-
-	t.Run("unsupport configstore", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockConfigStore := mock.NewMockStore(ctrl)
-		api := NewAPI("", nil, map[string]configstores.Store{"mock": mockConfigStore}, nil, nil, nil, nil, nil, nil, nil, nil)
-		_, err := api.SaveConfiguration(context.Background(), &runtimev1pb.SaveConfigurationRequest{StoreName: "etcd"})
-		assert.Equal(t, err.Error(), "configure store [etcd] don't support now")
-	})
-
-}
-
-func TestDeleteConfiguration(t *testing.T) {
-	t.Run("normal", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockConfigStore := mock.NewMockStore(ctrl)
-		mockConfigStore.EXPECT().Delete(gomock.Any(), gomock.Any()).DoAndReturn(func(ctx context.Context, req *configstores.DeleteRequest) error {
-			assert.Equal(t, "appid", req.AppId)
-			assert.Equal(t, 1, len(req.Keys))
-			assert.Equal(t, "key", req.Keys[0])
-			return nil
-		})
-		req := &runtimev1pb.DeleteConfigurationRequest{
-			StoreName: "mock",
-			AppId:     "appid",
-			Keys:      []string{"key"},
-			Metadata:  nil,
-		}
-		api := NewAPI("", nil, map[string]configstores.Store{"mock": mockConfigStore}, nil, nil, nil, nil, nil, nil, nil, nil)
-		_, err := api.DeleteConfiguration(context.Background(), req)
-		assert.Nil(t, err)
-	})
-
-	t.Run("unsupport configstore", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockConfigStore := mock.NewMockStore(ctrl)
-		api := NewAPI("", nil, map[string]configstores.Store{"mock": mockConfigStore}, nil, nil, nil, nil, nil, nil, nil, nil)
-		_, err := api.DeleteConfiguration(context.Background(), &runtimev1pb.DeleteConfigurationRequest{StoreName: "etcd"})
-		assert.Equal(t, err.Error(), "configure store [etcd] don't support now")
-	})
-
-}
-
-func TestSubscribeConfiguration(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockConfigStore := mock.NewMockStore(ctrl)
-	api := NewAPI("", nil, map[string]configstores.Store{"mock": mockConfigStore}, nil, nil, nil, nil, nil, nil, nil, nil)
-
-	//test not support store type
-	grpcServer := &MockGrpcServer{req: &runtimev1pb.SubscribeConfigurationRequest{}, err: nil}
-	err := api.SubscribeConfiguration(grpcServer)
-	assert.NotNil(t, err)
-	assert.Equal(t, err.Error(), "configure store [] don't support now")
-
-	//test
-	grpcServer2 := &MockGrpcServer{req: &runtimev1pb.SubscribeConfigurationRequest{}, err: errors.New("exit")}
-	err = api.SubscribeConfiguration(grpcServer2)
-	assert.NotNil(t, err)
-	assert.Equal(t, err.Error(), "exit")
-}
-
 type MockInvoker struct {
 	tmock.Mock
 }
@@ -314,394 +178,12 @@ func TestInvokeService(t *testing.T) {
 	})
 }
 
-func TestTryLock(t *testing.T) {
-	t.Run("lock store not configured", func(t *testing.T) {
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil, nil, nil)
-		req := &runtimev1pb.TryLockRequest{
-			StoreName: "abc",
-		}
-		_, err := api.TryLock(context.Background(), req)
-		assert.Equal(t, "rpc error: code = FailedPrecondition desc = lock store is not configured", err.Error())
-	})
-
-	t.Run("resourceid empty", func(t *testing.T) {
-		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
-		req := &runtimev1pb.TryLockRequest{
-			StoreName: "abc",
-		}
-		_, err := api.TryLock(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = ResourceId is empty in lock store abc", err.Error())
-	})
-
-	t.Run("lock owner empty", func(t *testing.T) {
-		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
-		req := &runtimev1pb.TryLockRequest{
-			StoreName:  "abc",
-			ResourceId: "resource",
-		}
-		_, err := api.TryLock(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = LockOwner is empty in lock store abc", err.Error())
-	})
-
-	t.Run("lock expire is not positive", func(t *testing.T) {
-		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
-		req := &runtimev1pb.TryLockRequest{
-			StoreName:  "abc",
-			ResourceId: "resource",
-			LockOwner:  "owner",
-		}
-		_, err := api.TryLock(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = Expire is not positive in lock store abc", err.Error())
-	})
-
-	t.Run("lock store not found", func(t *testing.T) {
-		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
-		req := &runtimev1pb.TryLockRequest{
-			StoreName:  "abc",
-			ResourceId: "resource",
-			LockOwner:  "owner",
-			Expire:     1,
-		}
-		_, err := api.TryLock(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = lock store abc not found", err.Error())
-	})
-
-	t.Run("normal", func(t *testing.T) {
-		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
-		mockLockStore.EXPECT().TryLock(gomock.Any()).DoAndReturn(func(req *lock.TryLockRequest) (*lock.TryLockResponse, error) {
-			assert.Equal(t, "lock|||resource", req.ResourceId)
-			assert.Equal(t, "owner", req.LockOwner)
-			assert.Equal(t, int32(1), req.Expire)
-			return &lock.TryLockResponse{
-				Success: true,
-			}, nil
-		})
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
-		req := &runtimev1pb.TryLockRequest{
-			StoreName:  "mock",
-			ResourceId: "resource",
-			LockOwner:  "owner",
-			Expire:     1,
-		}
-		resp, err := api.TryLock(context.Background(), req)
-		assert.Nil(t, err)
-		assert.Equal(t, true, resp.Success)
-	})
-
-}
-
-func TestUnlock(t *testing.T) {
-	t.Run("lock store not configured", func(t *testing.T) {
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil, nil, nil)
-		req := &runtimev1pb.UnlockRequest{
-			StoreName: "abc",
-		}
-		_, err := api.Unlock(context.Background(), req)
-		assert.Equal(t, "rpc error: code = FailedPrecondition desc = lock store is not configured", err.Error())
-	})
-
-	t.Run("resourceid empty", func(t *testing.T) {
-		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
-		req := &runtimev1pb.UnlockRequest{
-			StoreName: "abc",
-		}
-		_, err := api.Unlock(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = ResourceId is empty in lock store abc", err.Error())
-	})
-
-	t.Run("lock owner empty", func(t *testing.T) {
-		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
-		req := &runtimev1pb.UnlockRequest{
-			StoreName:  "abc",
-			ResourceId: "resource",
-		}
-		_, err := api.Unlock(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = LockOwner is empty in lock store abc", err.Error())
-	})
-
-	t.Run("lock store not found", func(t *testing.T) {
-		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
-		req := &runtimev1pb.UnlockRequest{
-			StoreName:  "abc",
-			ResourceId: "resource",
-			LockOwner:  "owner",
-		}
-		_, err := api.Unlock(context.Background(), req)
-		assert.Equal(t, "rpc error: code = InvalidArgument desc = lock store abc not found", err.Error())
-	})
-
-	t.Run("normal", func(t *testing.T) {
-		mockLockStore := mock_lock.NewMockLockStore(gomock.NewController(t))
-		mockLockStore.EXPECT().Unlock(gomock.Any()).DoAndReturn(func(req *lock.UnlockRequest) (*lock.UnlockResponse, error) {
-			assert.Equal(t, "lock|||resource", req.ResourceId)
-			assert.Equal(t, "owner", req.LockOwner)
-			return &lock.UnlockResponse{
-				Status: lock.SUCCESS,
-			}, nil
-		})
-		api := NewAPI("", nil, nil, nil, nil, nil, nil, map[string]lock.LockStore{"mock": mockLockStore}, nil, nil, nil)
-		req := &runtimev1pb.UnlockRequest{
-			StoreName:  "mock",
-			ResourceId: "resource",
-			LockOwner:  "owner",
-		}
-		resp, err := api.Unlock(context.Background(), req)
-		assert.Nil(t, err)
-		assert.Equal(t, runtimev1pb.UnlockResponse_SUCCESS, resp.Status)
-	})
-}
-
-func SendData(w net.Conn) {
-	w.Write([]byte("testFile"))
-	w.Close()
-}
-func TestGetFile(t *testing.T) {
-	r, w := net.Pipe()
-	ctrl := gomock.NewController(t)
-	mockFile := mock.NewMockFile(ctrl)
-	mockStream := runtime.NewMockRuntime_GetFileServer(ctrl)
-	api := NewAPI("", nil, nil, nil, nil, nil, map[string]file.File{"mock": mockFile}, nil, nil, nil, nil)
-	err := api.GetFile(&runtimev1pb.GetFileRequest{StoreName: "mock1"}, mockStream)
-	assert.Equal(t, err, status.Errorf(codes.InvalidArgument, "not supported store type: mock1"))
-	metadata := make(map[string]string)
-	mockFile.EXPECT().Get(context.Background(), &file.GetFileStu{FileName: "", Metadata: metadata}).Return(r, nil).Times(1)
-	mockStream.EXPECT().Send(&runtimev1pb.GetFileResponse{Data: []byte("testFile")}).Times(1)
-	mockStream.EXPECT().Context().Return(context.Background())
-	go SendData(w)
-	api.GetFile(&runtimev1pb.GetFileRequest{StoreName: "mock"}, mockStream)
-}
-
-func TestPutFile(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	mockFile := mock.NewMockFile(ctrl)
-	mockStream := runtime.NewMockRuntime_PutFileServer(ctrl)
-	api := NewAPI("", nil, nil, nil, nil, nil, map[string]file.File{"mock": mockFile}, nil, nil, nil, nil)
-
-	mockStream.EXPECT().Recv().Return(nil, io.EOF).Times(1)
-	err := api.PutFile(mockStream)
-	assert.Nil(t, err)
-
-	mockStream.EXPECT().Recv().Return(&runtimev1pb.PutFileRequest{StoreName: "mock1"}, nil).Times(1)
-	err = api.PutFile(mockStream)
-	assert.Equal(t, err, status.Errorf(codes.InvalidArgument, "not support store type: mock1"))
-
-	mockStream.EXPECT().Recv().Return(&runtimev1pb.PutFileRequest{StoreName: "mock"}, nil).Times(1)
-	stream := newPutObjectStreamReader(nil, mockStream)
-	Metadata := make(map[string]string)
-	mockStream.EXPECT().Context().Return(context.Background())
-	mockFile.EXPECT().Put(context.Background(), &file.PutFileStu{DataStream: stream, FileName: "", Metadata: Metadata}).Return(errors.New("err occur")).Times(1)
-	err = api.PutFile(mockStream)
-	s, _ := status.FromError(err)
-	assert.Equal(t, s.Message(), "err occur")
-}
-
-func TestListFile(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	mockFile := mock.NewMockFile(ctrl)
-	api := NewAPI("", nil, nil, nil, nil, nil, map[string]file.File{"mock": mockFile}, nil, nil, nil, nil)
-	request := &runtimev1pb.FileRequest{StoreName: "mock1"}
-	request.Metadata = make(map[string]string)
-	resp, err := api.ListFile(context.Background(), &runtimev1pb.ListFileRequest{Request: request})
-	assert.Nil(t, resp)
-	assert.Equal(t, err, status.Errorf(codes.InvalidArgument, "not support store type: mock1"))
-	request = &runtimev1pb.FileRequest{StoreName: "mock", Name: "test"}
-	request.Metadata = make(map[string]string)
-	mockFile.EXPECT().List(context.Background(), &file.ListRequest{DirectoryName: request.Name, Metadata: request.Metadata}).Return(&file.ListResp{Files: nil, Marker: "hello", IsTruncated: true}, nil).Times(1)
-	resp, err = api.ListFile(context.Background(), &runtimev1pb.ListFileRequest{Request: request})
-	assert.Nil(t, err)
-	assert.Equal(t, resp.Marker, "hello")
-	assert.Equal(t, resp.IsTruncated, true)
-	mockFile.EXPECT().List(context.Background(), &file.ListRequest{DirectoryName: request.Name, Metadata: request.Metadata}).Return(&file.ListResp{}, errors.New("test fail")).Times(1)
-	resp, err = api.ListFile(context.Background(), &runtimev1pb.ListFileRequest{Request: request})
-	assert.NotNil(t, err)
-	info := &file.FilesInfo{FileName: "hello", Size: 10, LastModified: "2021.11.12"}
-	files := make([]*file.FilesInfo, 0)
-	files = append(files, info)
-	mockFile.EXPECT().List(context.Background(), &file.ListRequest{DirectoryName: request.Name, Metadata: request.Metadata}).Return(&file.ListResp{Files: files}, nil).Times(1)
-	resp, err = api.ListFile(context.Background(), &runtimev1pb.ListFileRequest{Request: request})
-	assert.Equal(t, len(resp.Files), 1)
-	assert.Equal(t, resp.Files[0].FileName, "hello")
-}
-
-func TestDelFile(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	mockFile := mock.NewMockFile(ctrl)
-	api := NewAPI("", nil, nil, nil, nil, nil, map[string]file.File{"mock": mockFile}, nil, nil, nil, nil)
-	request := &runtimev1pb.FileRequest{StoreName: "mock1"}
-	request.Metadata = make(map[string]string)
-	resp, err := api.DelFile(context.Background(), &runtimev1pb.DelFileRequest{Request: request})
-	assert.Nil(t, resp)
-	assert.Equal(t, err, status.Errorf(codes.InvalidArgument, "not support store type: mock1"))
-	request = &runtimev1pb.FileRequest{StoreName: "mock", Name: "test"}
-	request.Metadata = make(map[string]string)
-	mockFile.EXPECT().Del(context.Background(), &file.DelRequest{FileName: request.Name, Metadata: request.Metadata}).Return(nil).Times(1)
-	_, err = api.DelFile(context.Background(), &runtimev1pb.DelFileRequest{Request: request})
-	assert.Nil(t, err)
-	mockFile.EXPECT().Del(context.Background(), &file.DelRequest{FileName: request.Name, Metadata: request.Metadata}).Return(errors.New("test fail")).Times(1)
-	_, err = api.DelFile(context.Background(), &runtimev1pb.DelFileRequest{Request: request})
-	assert.NotNil(t, err)
-}
-
-func TestGetFileMeta(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	mockFile := mock.NewMockFile(ctrl)
-	api := NewAPI("", nil, nil, nil, nil, nil, map[string]file.File{"mock": mockFile}, nil, nil, nil, nil)
-	request := &runtimev1pb.GetFileMetaRequest{Request: nil}
-	resp, err := api.GetFileMeta(context.Background(), request)
-	assert.Nil(t, resp)
-	st, _ := status.FromError(err)
-	assert.Equal(t, st.Message(), "request can't be nil")
-	request.Request = &runtimev1pb.FileRequest{StoreName: "mock", Name: "test"}
-	meta := make(map[string]string)
-	re := &file.FileMetaResp{
-		Size:         10,
-		LastModified: "123",
-		Metadata: map[string][]string{
-			"test": []string{},
-		},
-	}
-	mockFile.EXPECT().Stat(context.Background(), &file.FileMetaRequest{FileName: request.Request.Name, Metadata: meta}).Return(re, nil).Times(1)
-	resp, err = api.GetFileMeta(context.Background(), request)
-	assert.Equal(t, resp.LastModified, "123")
-	assert.Equal(t, int(resp.Size), 10)
-}
-
 func createTestClient(port int) *grpc.ClientConn {
 	conn, err := grpc.Dial(fmt.Sprintf("localhost:%d", port), grpc.WithInsecure())
 	if err != nil {
 		panic(err)
 	}
 	return conn
-}
-
-func TestGetSecret(t *testing.T) {
-	fakeStore := moke_secret.FakeSecretStore{}
-	fakeStores := map[string]secretstores.SecretStore{
-		"store1": fakeStore,
-		"store2": fakeStore,
-	}
-	testCases := []struct {
-		testName         string
-		storeName        string
-		key              string
-		errorExcepted    bool
-		expectedResponse string
-		expectedError    codes.Code
-	}{
-		{
-			testName:         "Good Key from store",
-			storeName:        "store1",
-			key:              "good-key",
-			errorExcepted:    false,
-			expectedResponse: "life is good",
-			expectedError:    codes.OK,
-		},
-		{
-			testName:         "error occur with error-key",
-			storeName:        "store2",
-			key:              "error-key",
-			errorExcepted:    true,
-			expectedResponse: "null",
-			expectedError:    codes.Internal,
-		},
-	}
-	// Setup API server
-	fakeAPI := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil,
-		nil, fakeStores)
-
-	// Run test server
-	port, _ := freeport.GetFreePort()
-	server := startTestRuntimeAPIServer(port, fakeAPI)
-	defer server.Stop()
-
-	// Create gRPC test client
-	clientConn := createTestClient(port)
-	defer clientConn.Close()
-	// act
-	client := runtimev1pb.NewRuntimeClient(clientConn)
-	for _, tt := range testCases {
-		t.Run(tt.testName, func(t *testing.T) {
-			req := &runtimev1pb.GetSecretRequest{
-				StoreName: tt.storeName,
-				Key:       tt.key,
-			}
-			resp, err := client.GetSecret(context.Background(), req)
-
-			if !tt.errorExcepted {
-				assert.NoError(t, err, "Expected no error")
-				assert.Equal(t, resp.Data[tt.key], tt.expectedResponse, "Expected responses to be same")
-			} else {
-				assert.Error(t, err, "Expected error")
-				assert.Equal(t, tt.expectedError, status.Code(err))
-			}
-		})
-	}
-}
-
-func TestGetBulkSecret(t *testing.T) {
-
-	fakeStore := moke_secret.FakeSecretStore{}
-	fakeStores := map[string]secretstores.SecretStore{
-		"store1": fakeStore,
-	}
-	expectedResponse := "life is good"
-
-	testCases := []struct {
-		testName         string
-		storeName        string
-		key              string
-		errorExcepted    bool
-		expectedResponse string
-		expectedError    codes.Code
-	}{
-		{
-			testName:         "Good Key from unrestricted store",
-			storeName:        "store1",
-			key:              "good-key",
-			errorExcepted:    false,
-			expectedResponse: expectedResponse,
-		},
-	}
-	// Setup API server
-	fakeAPI := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil,
-		nil, fakeStores)
-
-	// Run test server
-	port, _ := freeport.GetFreePort()
-	server := startTestRuntimeAPIServer(port, fakeAPI)
-	defer server.Stop()
-
-	// Create gRPC test client
-	clientConn := createTestClient(port)
-	defer clientConn.Close()
-
-	// act
-	client := runtimev1pb.NewRuntimeClient(clientConn)
-
-	for _, tt := range testCases {
-		t.Run(tt.testName, func(t *testing.T) {
-			req := &runtimev1pb.GetBulkSecretRequest{
-				StoreName: tt.storeName,
-			}
-			resp, err := client.GetBulkSecret(context.Background(), req)
-
-			if !tt.errorExcepted {
-				assert.NoError(t, err, "Expected no error")
-				assert.Equal(t, resp.Data[tt.key].Secrets[tt.key], tt.expectedResponse, "Expected responses to be same")
-			} else {
-				assert.Error(t, err, "Expected error")
-				assert.Equal(t, tt.expectedError, status.Code(err))
-			}
-		})
-	}
-
 }
 
 func TestNewGrpcServer(t *testing.T) {
@@ -711,28 +193,6 @@ func TestNewGrpcServer(t *testing.T) {
 		t.Error()
 		return
 	}
-}
-
-func TestInvokeBinding(t *testing.T) {
-	port, _ := freeport.GetFreePort()
-	srv := NewAPI("", nil, nil, nil, nil, nil, nil, nil, nil,
-		func(name string, req *bindings.InvokeRequest) (*bindings.InvokeResponse, error) {
-			if name == "error-binding" {
-				return nil, errors.New("error when invoke binding")
-			}
-			return &bindings.InvokeResponse{Data: []byte("ok")}, nil
-		}, nil)
-	server := startTestServerAPI(port, srv)
-	defer server.Stop()
-
-	clientConn := createTestClient(port)
-	defer clientConn.Close()
-
-	client := runtimev1pb.NewRuntimeClient(clientConn)
-	_, err := client.InvokeBinding(context.Background(), &runtimev1pb.InvokeBindingRequest{})
-	assert.Nil(t, err)
-	_, err = client.InvokeBinding(context.Background(), &runtimev1pb.InvokeBindingRequest{Name: "error-binding"})
-	assert.Equal(t, codes.Internal, status.Code(err))
 }
 
 func startTestServerAPI(port int, srv runtimev1pb.RuntimeServer) *grpc.Server {
